requirements.txt
Flask==3.1.2
Werkzeug==3.1.3
python-dotenv==1.0.1
supabase==2.24.0

